feat:AGED-231:
- Edit mission screen
- Use gson adapter only
- Move image creation to ImageRepository
- Create Outbound and Inboud mission

+++ mission/data/src/main/java/com/upsaclay/mission/data/local/LocalMission.kt
@@ -24,7 +24,7 @@ const val MISSION_TABLE = "missions_table"
 data class LocalMission(
     @PrimaryKey
     @ColumnInfo(name = MISSION_ID)
-    val missionId: Int,
+    val missionId: Long,
     @ColumnInfo(name = TITLE)
     val missionTitle: String,
     @ColumnInfo(name = DESCRIPTION)
